requirements.txt
flask==2.3.3
numpy==1.23.5
opencv-python-headless==4.8.0.76
tensorflow==2.12.0
pillow==10.0.0
scikit-image==0.21.0
scikit-learn==1.3.0
matplotlib==3.7.2
gunicorn==21.2.0
Flask-SocketIO==5.3.4
eventlet==0.33.3